+++ webauthn_sdk/src/lib.rs
@@ -1,7 +1,2 @@
-extern crate termcolor;
-
-pub mod compile;
-pub mod generate;
-pub mod logger;
-pub mod parse;
+pub mod prepare;
 pub mod run;
